deletion route for courses functional
