+++ controllers/Courses.js
@@ -4,6 +4,7 @@ const { deSerializeCourse } = require('../serializers/Course');
 const Course = require('../models/Course');
 const HttpError = require('http-errors');
 
+
 /*
 async function deleteCourse(sessionToken, Id) {
   const request = axios.create({
@@ -30,7 +31,23 @@ async function findOne(sessionToken, criteria) {
   }
 }
 */
-
+async function deleteCourse(sessionToken, courseId) {
+  const request = axios.create({
+    headers: { Authorization: `Bearer ${sessionToken}` },
+  });
+  const response = await request.delete(`courses/${courseId}`);
+  if(response.status === 200) {
+    return {
+      message: 'Course was deleted successfully',
+      status: request.status
+    };
+  }
+  log.debug(`There was an error deleting course with status code: ${response.status}`);
+  return {
+    message: response.message,
+    status: request.status
+  };
+}
 async function findAll(sessionToken, criteria, limit = 100, offset = 0) {
   const request = axios.create({
     headers: { Authorization: `Bearer ${sessionToken}` },
@@ -70,8 +87,8 @@ async function createCourse(sessionToken, requestBody) {
 }
 
 module.exports = {
-  // deleteCourse,
   // findOne,
+  deleteCourse,
   findAll,
   createCourse,
 };
