+++ controllers/Courses.test.js
@@ -16,6 +16,22 @@ describe('Course controller tests', () => {
     axios.post.mockReset();
   });
 
+  describe('deleteCourses tests', () => {
+    test('if deletion of course was successful', async () => {
+      const course = 
+        {
+          id: 1,
+          courseId: 157,
+          name: 'Mathematics',
+          credits: 3,
+          section: 1
+        };
+      axios.post.mockResolvedValueOnce({ status: 201 });
+      const result = await Course.deleteCourse('session-token', course);
+      expect(result.message).toEqual('Course was deleted successfully');
+    });
+  });
+
 
   describe('fetchAll tests', () => {
     test('fetchAll - happy path test', async () => {
