+++ static/css/style.css
@@ -204,7 +204,9 @@ body {
   opacity: 90%;
   color: #fff;
 }
-
+.modal-title {
+  color: #0c233f;
+}
 .role-alert-location {
   position: absolute;
   top: 50px;
