+++ views/pages/courses/index.ejs
@@ -94,35 +94,20 @@ futureCourses = [{
         <a class="nav-link">Courses</a>
       </li>
       <li class="nav-pill ml-auto">
-        <button id="create-course-button" type="button" class="btn nav-link active" data-toggle="modal"
+        <button id="create-course-button" type="button" class="btn nav-link btn-stout" data-toggle="modal"
           data-target="#create-course-modal">
           Create Course
         </button>
       </li>
     </ul>
     <hr>
-    <button type="button" class="btn btn-primary" onClick="showCompleted()" id="completeBtn">Completed Courses</button>
+    <button type="button" class="btn btn-primary btn-stout" onclick="showCompleted()" id="completeBtn">Completed
+      Courses</button>
 
-    <button type="button" class="btn btn-primary" onclick="showCurrent()" id="currBtn">Current Courses</button>
+    <button type="button" class="btn btn-primary btn-stout" onclick="showCurrent()" id="currBtn">Current
+      Courses</button>
 
-    <button type="button" class="btn btn-primary" onclick="viewCourses()" id="viewBtn">
-        <button
-        id="create-course-button"
-        type="button"
-        class="btn nav-link btn-stout"
-        data-toggle="modal"
-        data-target="#create-course-modal"
-        >
-        Create Course
-      </button>
-    </li>
-  </ul>
-  <hr>
-  <button type="button" class="btn btn-primary btn-stout" onClick="showCompleted()" id="completeBtn">Completed Courses</button>
-  
-  <button type="button" class="btn btn-primary btn-stout" onclick="showCurrent()" id="currBtn">Current Courses</button>
-  
-  <button type="button" class="btn btn-primary btn-stout" onclick="viewCourses()" id="viewBtn">
+    <button type="button" class="btn btn-primary btn-stout" onclick="viewCourses()" id="viewBtn">
       View Courses
     </button>
   </div>
@@ -269,6 +254,9 @@ futureCourses = [{
                     Not Defined Yet
                   </ul>
                 </ul>
+              </div>
+              <div class="modal-footer">
+                <button type="button" class="btn btn-danger">Delete Course</button>
               </div>
             </div>
           </div>
